+++ src/main/java/racingcar/constant/GameMessage.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.constant;
 
 public enum GameMessage {
     START_GAME("경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)"),
