+++ src/main/java/racingcar/view/InputValidator.java
@@ -1,6 +1,6 @@
-package racingcar;
+package racingcar.view;
 
-import java.util.regex.Matcher;
+import racingcar.constant.ExceptionMessage;
 
 public class InputValidator {
